Move the unit tests to separate file in "TokenWrapper" contract
