+++ contracts/tokenwrapper/src/lib.rs
@@ -3,3 +3,5 @@ pub mod state;
 
 #[cfg(test)]
 pub mod mock_querier;
+#[cfg(test)]
+pub mod tests;
